requirements.txt
matplotlib==3.1.2
numpy==1.17.4
opencv_python==4.6.0.66
Pillow==9.4.0
scikit_learn==1.2.0
seaborn==0.11.2
torch==1.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.62.3